Apply PR #5 review fixes for ticketing
Backend (functions/src/index.ts):
- archiveUpcomingEvent Phase C re-queries live subcollections so docs written between Phase A and B no longer strand under the deleted event
- sendTicketEmails enforces the Resend daily cap server-side and marks zero-ticketId attendees as sent to prevent infinite drain
- Unsent drain orders by createdAt for FIFO fairness (uses existing index)
- ticket-attendee-edit audit record written on name-only attendee edits
- QR PNG width drops from 400 to 256 to keep mail docs under Firestore's 1 MB limit even for high-ticket attendees
- getTicketEmailQuota returns chunkSize so the client doesn't duplicate it

Frontend:
- Split TicketsSubtab.tsx (1552 LOC) into app/pages/admin/tickets/
  {TicketsSubtab,AttendeesSection,ImportSection,TemplateSection, SendSection,helpers,types}
- New shared sanitize allowlist at app/lib/emailSanitize.ts; template preview now sanitizes with the same options the server applies on save
- claim.tsx: ticket-flow not-logged-in / not-authorized states now read "Show This at the Door" so attendees opening their own QR aren't confused; sign-in button hidden in that flow
- RecordsTab + admin types render the new ticket-attendee-edit type
- DEPLOY.md documents why the root package-lock.json is intentionally uncommitted (functions/package-lock.json stays committed)
