+++ package.json
@@ -24,7 +24,8 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
-    "react-router": "^7.5.3"
+    "react-router": "^7.5.3",
+    "sanitize-html": "^2.13.0"
   },
   "devDependencies": {
     "@react-router/dev": "^7.8.2",
@@ -33,6 +34,7 @@
     "@types/papaparse": "^5.3.15",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
+    "@types/sanitize-html": "^2.13.0",
     "tailwindcss": "^4.1.4",
     "typescript": "^5.8.3",
     "vite": "^6.3.3",
